+++ R/diffnet-methods.R
@@ -895,7 +895,7 @@ plot_infectsuscep.list <- function(graph, toa, t0, normalize,
 #' @examples
 #' # Generating a random diffnet -----------------------------------------------
 #' set.seed(821)
-#' diffnet <- rdiffnet(20, 5, seed.graph="small-world", seed.nodes="central")
+#' diffnet <- rdiffnet(100, 5, seed.graph="small-world", seed.nodes="central")
 #'
 #' plot_adopters(diffnet)
 #'