+++ src/netdiffuser_extra.cpp
@@ -12,17 +12,17 @@ arma::umat sparse_indexes(const arma::sp_mat & mat) {
   // If the matrix is empty (which makes no sense)
   if (!n) return indices;
 
-  int curcol = 1;
-  for (int i=0;i<n;i++) {
+  // More efficient implementation
+  arma::sp_mat::const_iterator begin = mat.begin();
+  arma::sp_mat::const_iterator end   = mat.end();
 
-    // Incrementing column while there's no new element in the col
-    while (mat.col_ptrs[curcol] <= i) ++curcol;
-    // while (mat.col_ptrs[curcol-1] == mat.col_ptrs[curcol]) ++curcol;
+  int i = 0;
+  for (arma::sp_mat::const_iterator it = begin; it != end; ++it) {
+    indices.at(i, 0) = it.row();
+    indices.at(i, 1) = it.col();
 
-    // Asigning indexes
-    indices.at(i, 0) = mat.row_indices[i];
-    // indices.at(i,1) = j;
-    indices.at(i, 1) = curcol - 1;
+    // Incrementing
+    i++;
   }
   // return indices;
   return indices;
@@ -44,4 +44,37 @@ double angle(double x0, double y0, double x1, double y1) {
   else if ((xdist < 0) && (ydist < 0)) return(alpha + PI);
 
   return(alpha);
+}
+
+
+// [[Rcpp::export]]
+arma::sp_mat sp_trimatl(const arma::sp_mat & x) {
+  // Getting start-end
+  arma::sp_mat::const_iterator start = x.begin();
+  arma::sp_mat::const_iterator end   = x.end();
+
+  // Empty mat
+  int n = x.n_cols;
+  arma::sp_mat newx(n,n);
+
+  for(arma::sp_mat::const_iterator it = start; it != end; ++it) {
+    int i = it.row();
+    int j = it.col();
+    if (i >= j)
+      newx.at(i,j) = *it;
+  }
+
+  return newx;
+}
+
+// [[Rcpp::export]]
+arma::sp_mat sp_diag(const arma::sp_mat & x, const arma::vec & v) {
+  // Checking dimensions
+  if (x.n_cols != x.n_rows) Rcpp::stop("-x- must be square.");
+  if (x.n_cols != v.n_elem) Rcpp::stop("length(v) must be equal to ncol(x)");
+
+  arma::sp_mat out(x);
+  out.diag() = v;
+
+  return out;
 }