+++ src/rgraph_ws.h
@@ -9,7 +9,7 @@ using namespace Rcpp;
 
 arma::sp_mat ring_lattice(int n, int k, bool undirected=true);
 
-arma::sp_mat rewire_graph_cpp(
+arma::sp_mat rewire_endpoints(
     const arma::sp_mat & graph, double p,
     bool both_ends=false, bool self=false, bool multiple=false,
     bool undirected=false);