+++ tests/testthat/test-random_graph.R
@@ -141,3 +141,38 @@ test_that("When p=1 in rewiring, Pr(j'=i) = Pr(j'=k) for all (i,k) in V", {
   # plot(m-1/n, type="l", ylim=c(-.00025,.00025))
   expect_equal(m - 1/(n), x, tolerance=.00025, check.attributes=FALSE)
 })
+
+# Rewiring degree preserve
+test_that("rewire_graph_const_cpp should hold degree", {
+  set.seed(18231)
+  n <- 1e3
+  N <- 1e2
+
+  # Function to compute degrees
+  dfun <- function(x) cbind(dgr(x, "indegree"), dgr(x, "outdegree"))
+
+  # Directed graph
+  out <- vector(length = n)
+  for (i in 1:n) {
+    x  <- rgraph_ba(t=N-1)
+    x  <- netdiffuseR:::sp_diag(x, rep(0, N))
+    d0 <- dfun(x)
+    y  <- netdiffuseR:::rewire_swap(x, 100)
+    d1 <- dfun(y)
+
+    out[i] <- identical(d0, d1)
+  }
+  expect_equal(out, rep(TRUE, n))
+
+  # Undirected graph
+  out <- vector(length = n)
+  for (i in 1:n) {
+    x  <- rgraph_ws(n=N-1, k=4, p=.3)
+    d0 <- dfun(x)
+    y  <- netdiffuseR:::rewire_swap(x, 100, undirected = TRUE)
+    d1 <- dfun(y)
+
+    out[i] <- identical(d0, d1)
+  }
+  expect_equal(out, rep(TRUE, n))
+})